requirements.txt
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
numpy==1.24.2
pandas==1.5.3
gunicorn==21.2.0
# Django==4.2.3
azure-ai-formrecognizer
azure-core
pdfkit==1.0.0
google-api-core==2.16.1
google-api-python-client==2.116.0
google-auth==2.23.4
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.61.0
pdfkit==1.0.0
docx2pdf==0.1.8
pywin32==306
docxtpl==0.17.0
pdf2image==1.17.0
pytesseract==0.3.10
